[MODIFY] change the data type of heuristic_dist from int to double

+++ inc/a_star_node.h
@@ -7,13 +7,13 @@ struct node_heap_t
 {
     int vertexName;
     int real_dist;
-    int total_dist_to_src;
+    double total_dist_to_src;
     struct node_heap_t *next;
 };
 
 typedef struct node_heap_t node_heap_t;
 
-node_heap_t *NODE_HEAP_Create(int vertex, int dist, int heuristic_dist);
+node_heap_t *NODE_HEAP_Create(int vertex, double dist, double heuristic_dist);
 void NODE_HEAP_Free(node_heap_t *p_node_heap);
 
 #endif
